initial 7.0 updates

+++ VFXEditor/Select/Tabs/Statuses/StatusRow.cs
@@ -16,7 +16,7 @@ namespace VfxEditor.Select.Tabs.Statuses {
         public StatusRow( Status status ) {
             Name = status.Name.ToString();
             RowId = ( int )status.RowId;
-            Icon = status.Icon;
+            Icon = ( uint )status.Icon;
 
             HitPath = GetVfxPath( status.HitEffect.Value?.Location.Value?.Location );
